Big update:
- For refreshing the screen after spacing is changed only drawTablature
needs to be called.  Not createTab.  

- implemented an algorithm which concatenates all the staves into
one large stave.  Then partitions the big stave into sets of measures
whose cardinality is dependent on the value which the spacing is set
to.

- re-implemented the drawing procedure to use regular expressions to
detect patterns within tokens with a length of 5.

- Did some re-factoring, but more can be done, I may move all of the
variables and methods which reorganize the measures according to
spacing to the Tablature class.  I may keep all methods in they're own
class just for drawing.

BUGS THAT I STILL NEED TO KILL:
1. Implement a method which cuts out all white space characters
2.long measures with a spacing more than 11 causes the system to hang.
(workaround, limit spacing to under 11 until it's fixed)
3. The little "h" above the hammer curve has to be drawn still.
4. Implement the pull off symbol.   eg: 0p3 

COMMENTS:
Right now it is good enough to add a GUI too, I will try and get the
above 4 things fixed, especially the drawing methods.  

I also need to add some more comments.  But outside this class the code
is the same, so there should not be any problems adding a GUI and other
functionality. I still have to make unit tests for the little helper
methods, but the major methods have unit tests.

 Roughly 800 lines of new code were added to make it
dynamic, so it might be hard to understand just by looking at it.   On
Monday I will try and have some  diagrams to show how it works exactly
so we will all be able to answer the professors questions during the
mid-term presentation.